imagens

+++ backend/src/main/java/br/unitins/tp1/pizzaria/resource/ProdutoResource.java
@@ -74,7 +74,7 @@ public class ProdutoResource {
         String nomeImagem;
         try {
             String old = service.findById(id).nomeImagem();
-            if(!old.isBlank()){
+            if(old != null && !old.isBlank()){
                 imageService.remover(old);
             }
             nomeImagem = imageService.salvar(form.getNomeImagem(), form.getImagem());
